Move TrimVariableGlyphs to glyph.py, since it will have code in common with DecomposeComposites
